Add server name hash config
The server_names_hash_max_size is usually configured according to the CPU. However we reached our maximum number of sever names so we needed to manually override this. Extra care will have to be taken in future Trellis updates so as not to remove this setting. This is also a quick fix, and not final. Performance is going to take a hit with this setting in place so this number needs bringing down as much as possible in a future commit

Ref: http://nginx.org/en/docs/http/server_names.html#optimization