Disable X-Content-Type-Options
At the moment there is issues with webpack injecting styles with incorrect header types on local watch commands. This is a temporary work around whilst the issue is resolved.

In future we may consider conditionally disabling this for development envs and make this an Ansible managed file. I’m closely monitoring this in the discourse and GH issue